mcp_server/requirements.txt
fastmcp>=0.1.0
boto3>=1.28.38
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0